Add small section on async/await
We plan to add a study on these concepts, but many instructors already
show off async/await during this lesson. Added it as an item to "refactor"
in the code along for the advanced react section. Also added a bonus lab
if there is extra time to have folks continue to change the remaining
Promise chains into async/await functions.

Solution is in the NOTES.md file as well.

References #8